fix: unnecessary int cast on gift cards, added test for cents not just whole dollars

+++ src/dao/gift-card.dao.ts
@@ -180,7 +180,7 @@ const currentBalanceExpr = `
       WHEN 'balance_redeemed' THEN -e.amount
       ELSE 0
     END
-  ), 0)::int AS current_balance
+  ), 0) AS current_balance
 `;
 
 function mapDbRowToGiftCard(row: GiftCard): GiftCard {
